projects/export/pyproject.toml
[tool.poetry]
name = "export"
version = "0.0.1"
description = "Export DeepClean for IaaS inference"
authors = ["Alec Gunny <alec.gunny@gmail.com>"]
# readme = "README.md"

[tool.poetry.scripts]
train = "export:main"

[tool.poetry.dependencies]
python = "^3.8,<3.10"

"deepclean.networks" = {path = "../../libs/network", develop = true}
"deepclean.logging" = {path = "../../libs/logging", develop = true}

"hermes.quiver" = {path = "../../gw-iaas/libs/hermes/hermes.quiver", develop = true}
"hermes.typeo" = {path = "../../gw-iaas/libs/hermes/hermes.typeo", develop = true}

[tool.poetry.dev-dependencies]
pytest = "^6.2"

[build-system]
requires = ["poetry>=1.2"]
build-backend = "poetry.masonry.api"

[tool.typeo]
channels = [
    "H1:GDS-CALIB_STRAIN",
    "H1:PEM-CS_MAINSMON_EBAY_1_DQ",
    "H1:ASC-INP1_P_INMON",
    "H1:ASC-INP1_Y_INMON",
    "H1:ASC-MICH_P_INMON",
    "H1:ASC-MICH_Y_INMON",
    "H1:ASC-PRC1_P_INMON",
    "H1:ASC-PRC1_Y_INMON",
    "H1:ASC-PRC2_P_INMON",
    "H1:ASC-PRC2_Y_INMON",
    "H1:ASC-SRC1_P_INMON",
    "H1:ASC-SRC1_Y_INMON",
    "H1:ASC-SRC2_P_INMON",
    "H1:ASC-SRC2_Y_INMON",
    "H1:ASC-DHARD_P_INMON",
    "H1:ASC-DHARD_Y_INMON",
    "H1:ASC-CHARD_P_INMON",
    "H1:ASC-CHARD_Y_INMON",
    "H1:ASC-DSOFT_P_INMON",
    "H1:ASC-DSOFT_Y_INMON",
    "H1:ASC-CSOFT_P_INMON",
    "H1:ASC-CSOFT_Y_INMON"
]
repository_directory = "../../repo"
weights = "results/test"
kernel_length = 1
stride_length = 0.001953125
sample_rate = 4096
max_latency = 0.5
instances = 6
verbose = false

[tool.typeo.commands.autoencoder]
